+++ src/indexer/mavryk/handlers/on_doorman_unpause_all.py
@@ -19,7 +19,7 @@ async def on_doorman_unpause_all(
     doorman.unstake_paused                  = unpause_all.storage.breakGlassConfig.unstakeIsPaused
     doorman.compound_paused                 = unpause_all.storage.breakGlassConfig.compoundIsPaused
     doorman.farm_claim_paused               = unpause_all.storage.breakGlassConfig.farmClaimIsPaused
-    doorman.on_vault_deposit_smvk_paused    = unpause_all.storage.breakGlassConfig.onVaultDepositStakedMvkIsPaused
-    doorman.on_vault_withdraw_smvk_paused   = unpause_all.storage.breakGlassConfig.onVaultWithdrawStakedMvkIsPaused
-    doorman.on_vault_liquidate_smvk_paused  = unpause_all.storage.breakGlassConfig.onVaultLiquidateStakedMvkIsPaused
+    doorman.on_vault_deposit_stake_paused    = unpause_all.storage.breakGlassConfig.onVaultDepositStakeIsPaused
+    doorman.on_vault_withdraw_stake_paused   = unpause_all.storage.breakGlassConfig.onVaultWithdrawStakeIsPaused
+    doorman.on_vault_liquidate_stake_paused  = unpause_all.storage.breakGlassConfig.onVaultLiquidateStakeIsPaused
     await doorman.save()