+++ src/indexer/mavryk/handlers/on_lending_controller_set_collateral_token.py
@@ -20,13 +20,18 @@ async def on_lending_controller_set_collateral_token(
 
     lending_controller_address      = set_collateral_token.data.target_address
     if collateral_token_name in set_collateral_token.storage.collateralTokenLedger:
-        collateral_token_storage        = set_collateral_token.storage.collateralTokenLedger[collateral_token_name]
-        collateral_token_oracle_address = collateral_token_storage.oracleAddress
-        collateral_token_address        = collateral_token_storage.tokenContractAddress
-        collateral_token_protected      = collateral_token_storage.protected
-        collateral_token_scaled         = collateral_token_storage.isScaledToken
-        collateral_token_id             = 0
-        collateral_token_standard       = ""
+        collateral_token_storage                    = set_collateral_token.storage.collateralTokenLedger[collateral_token_name]
+        collateral_token_oracle_address             = collateral_token_storage.oracleAddress
+        collateral_token_address                    = collateral_token_storage.tokenContractAddress
+        collateral_token_protected                  = collateral_token_storage.protected
+        collateral_token_scaled                     = collateral_token_storage.isScaledToken
+        collateral_token_staked                     = collateral_token_storage.isStakedToken
+        collateral_token_staking_contract_address   = collateral_token_storage.stakingContractAddress
+        collateral_token_total_deposited            = float(collateral_token_storage.totalDeposited)
+        collateral_token_max_deposited_amount       = float(collateral_token_storage.maxDepositedAmount)
+        collateral_token_paused                     = collateral_token_storage.isPaused
+        collateral_token_id                         = 0
+        collateral_token_standard                   = ""
 
         # Persist collateral Token Metadata
         await persist_token_metadata(
@@ -62,8 +67,13 @@ async def on_lending_controller_set_collateral_token(
             token_address       = collateral_token_address,
             oracle              = oracle
         )
-        lending_controller_collateral_token.protected                   = collateral_token_protected
-        lending_controller_collateral_token.is_scaled_token             = collateral_token_scaled
-        lending_controller_collateral_token.token_name                  = collateral_token_name
-        lending_controller_collateral_token.token_contract_standard     = collateral_token_standard
+        lending_controller_collateral_token.protected                                   = collateral_token_protected
+        lending_controller_collateral_token.is_scaled_token                             = collateral_token_scaled
+        lending_controller_collateral_token.token_name                                  = collateral_token_name
+        lending_controller_collateral_token.token_contract_standard                     = collateral_token_standard
+        lending_controller_collateral_token.is_staked_token                             = collateral_token_staked
+        lending_controller_collateral_token.staking_contract_address                    = collateral_token_staking_contract_address
+        lending_controller_collateral_token.total_deposited                             = collateral_token_total_deposited
+        lending_controller_collateral_token.max_deposited_amount                        = collateral_token_max_deposited_amount
+        lending_controller_collateral_token.paused                                      = collateral_token_paused
         await lending_controller_collateral_token.save()