+++ src/indexer/mavryk/handlers/on_vault_factory_create_vault.py
@@ -1,11 +1,11 @@
 
 from dipdup.models import Origination
 from mavryk.utils.persisters import persist_contract_metadata
-from mavryk.types.vault_factory.parameter.create_vault import CreateVaultParameter
+from mavryk.types.vault_factory.parameter.create_vault import CreateVaultParameter, DepositorsConfigItem as Any, DepositorsConfigItem1 as Whitelist
 from dipdup.context import HandlerContext
 from dipdup.models import Transaction
 from mavryk.types.vault_factory.storage import VaultFactoryStorage
-from mavryk.types.vault.storage import VaultStorage, Depositor as Any, Depositor1 as Whitelist
+from mavryk.types.vault.storage import VaultStorage
 import mavryk.models as models
 
 async def on_vault_factory_create_vault(
@@ -23,9 +23,9 @@ async def on_vault_factory_create_vault(
     depositors              = vault_origination.storage.depositors
     allowance_type          = models.VaultAllowance.ANY
 
-    if type(depositors) == Any:
+    if type(depositors.depositorsConfig) == Any:
         allowance_type  = models.VaultAllowance.ANY
-    elif type(depositors) == Whitelist:
+    elif type(depositors.depositorsConfig) == Whitelist:
         allowance_type  = models.VaultAllowance.WHITELIST
 
     # Check vault does not already exists
@@ -74,8 +74,8 @@ async def on_vault_factory_create_vault(
         await vault.save()
 
         # Register depositors
-        if type(depositors) == Whitelist:
-            for depositor_address in depositors.whitelist:
+        if type(depositors.depositorsConfig) == Whitelist:
+            for depositor_address in depositors.whitelistedDepositors:
                 depositor           = await models.mavryk_user_cache.get(address=depositor_address)
                 vault_depositor, _  = await models.VaultDepositor.get_or_create(
                     vault       = vault,