+++ src/indexer/mavryk/sql_model/enums.py
@@ -5,74 +5,74 @@ from enum import IntEnum
 ###
 
 class DexType(IntEnum):
-    ADD_LIQUIDITY           = 0
-    REMOVE_LIQUIDITY        = 1
-    XTZ_TO_TOKEN            = 2
-    TOKEN_TO_XTZ            = 3
-    TOKEN_TO_TOKEN          = 4
+    ADD_LIQUIDITY                   = 0
+    REMOVE_LIQUIDITY                = 1
+    XTZ_TO_TOKEN                    = 2
+    TOKEN_TO_XTZ                    = 3
+    TOKEN_TO_TOKEN                  = 4
 
 class StakeType(IntEnum):
-    STAKE                   = 0
-    UNSTAKE                 = 1
-    FARM_CLAIM              = 2
-    COMPOUND                = 3
-    SATELLITE_REWARD        = 4
-    VAULT_DEPOSIT_SMVK      = 5
-    VAULT_WITHDRAW_SMVK     = 6
-    VAULT_LIQUIDATE_SMVK    = 7
+    STAKE                           = 0
+    UNSTAKE                         = 1
+    FARM_CLAIM                      = 2
+    COMPOUND                        = 3
+    SATELLITE_REWARD                = 4
+    VAULT_DEPOSIT_STAKED_TOKEN      = 5
+    VAULT_WITHDRAW_STAKED_TOKEN     = 6
+    VAULT_LIQUIDATE_STAKED_TOKEN    = 7
 
 class ActionStatus(IntEnum):
-    PENDING                 = 0
-    FLUSHED                 = 1
-    EXECUTED                = 2
+    PENDING                         = 0
+    FLUSHED                         = 1
+    EXECUTED                        = 2
 
 class SatelliteStatus(IntEnum):
-    ACTIVE                  = 0
-    SUSPENDED               = 1
-    BANNED                  = 2
+    ACTIVE                          = 0
+    SUSPENDED                       = 1
+    BANNED                          = 2
 
 class GovernanceRoundType(IntEnum):
-    PROPOSAL                = 0
-    VOTING                  = 1
-    TIMELOCK                = 2
+    PROPOSAL                        = 0
+    VOTING                          = 1
+    TIMELOCK                        = 2
 
 class GovernanceActionStatus(IntEnum):
-    ACTIVE                  = 0
-    DROPPED                 = 1
+    ACTIVE                          = 0
+    DROPPED                         = 1
 
 class GovernanceVoteType(IntEnum):
-    NAY                     = 0
-    YAY                     = 1
-    PASS                    = 2
+    NAY                             = 0
+    YAY                             = 1
+    PASS                            = 2
 
 class TokenType(IntEnum):
-    XTZ                     = 0
-    FA12                    = 1
-    FA2                     = 2
-    OTHER                   = 3
+    XTZ                             = 0
+    FA12                            = 1
+    FA2                             = 2
+    OTHER                           = 3
 
 class RewardType(IntEnum):
-    XTZ                     = 0
-    SMVK                    = 1
+    XTZ                             = 0
+    SMVK                            = 1
 
 class VaultAllowance(IntEnum):
-    ANY                     = 0
-    WHITELIST               = 1
+    ANY                             = 0
+    WHITELIST                       = 1
 
 class LendingControllerOperationType(IntEnum):
-    ADD_LIQUIDITY           = 0
-    REMOVE_LIQUIDITY        = 1
-    BORROW                  = 2
-    REPAY                   = 3
-    DEPOSIT                 = 4
-    WITHDRAW                = 5
-    DEPOSIT_SMVK            = 6
-    WITHDRAW_SMVK           = 7
-    VAULT_CREATION          = 8
-    MARK_FOR_LIQUIDATION    = 9
-    LIQUIDATE_VAULT         = 10
-    CLOSE_VAULT             = 11
+    ADD_LIQUIDITY                   = 0
+    REMOVE_LIQUIDITY                = 1
+    BORROW                          = 2
+    REPAY                           = 3
+    DEPOSIT                         = 4
+    WITHDRAW                        = 5
+    DEPOSIT_STAKED_TOKEN            = 6
+    WITHDRAW_STAKED_TOKEN           = 7
+    VAULT_CREATION                  = 8
+    MARK_FOR_LIQUIDATION            = 9
+    LIQUIDATE_VAULT                 = 10
+    CLOSE_VAULT                     = 11
 
 class MTokenOperationType(IntEnum):
-    TRANSFER                = 0
-    MINT_OR_BURN            = 1
+    TRANSFER                        = 0
+    MINT_OR_BURN                    = 1